Rename and move some things.
